requirements.txt
asgiref==3.8.1
colorthief==0.2.1
distlib==0.3.6
Django==5.2
djangorestframework==3.16.0
et-xmlfile==1.1.0
filelock==3.9.0
git-filter-repo==2.47.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.1.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.20.0

reportlab==4.2.2

pillow==10.4.0
